+++ src/main/java/amidst/PerApplicationInjector.java
@@ -19,8 +19,8 @@ import amidst.gui.profileselect.ProfileSelectWindow;
 import amidst.mojangapi.MojangApi;
 import amidst.mojangapi.MojangApiBuilder;
 import amidst.mojangapi.file.DotMinecraftDirectoryNotFoundException;
-import amidst.mojangapi.file.facade.PlayerInformationCache;
-import amidst.mojangapi.file.facade.PlayerInformationProvider;
+import amidst.mojangapi.file.PlayerInformationCache;
+import amidst.mojangapi.file.PlayerInformationProvider;
 import amidst.mojangapi.minecraftinterface.local.LocalMinecraftInterfaceCreationException;
 import amidst.mojangapi.world.SeedHistoryLogger;
 import amidst.mojangapi.world.World;
