+++ src/main/java/amidst/mojangapi/MojangApi.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 import amidst.documentation.NotNull;
 import amidst.documentation.ThreadSafe;
 import amidst.logging.AmidstLogger;
+import amidst.mojangapi.file.LauncherProfile;
+import amidst.mojangapi.file.MinecraftInstallation;
 import amidst.mojangapi.file.MojangApiParsingException;
-import amidst.mojangapi.file.facade.LauncherProfile;
-import amidst.mojangapi.file.facade.MinecraftInstallation;
-import amidst.mojangapi.file.facade.VersionList;
+import amidst.mojangapi.file.VersionList;
 import amidst.mojangapi.minecraftinterface.MinecraftInterface;
 import amidst.mojangapi.minecraftinterface.MinecraftInterfaceException;
 import amidst.mojangapi.minecraftinterface.RecognisedVersion;
