+++ src/main/java/amidst/mojangapi/file/ImmutablePlayerInformationProvider.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.file.facade;
+package amidst.mojangapi.file;
 
 import amidst.documentation.Immutable;
 import amidst.documentation.NotNull;
