+++ src/main/java/amidst/mojangapi/file/SaveGame.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.file.facade;
+package amidst.mojangapi.file;
 
 import java.util.Collections;
 import java.util.LinkedList;
